+++ src/app/routers/files.py
@@ -6,7 +6,6 @@ from uuid import uuid4
 from app.schemas.file_schemas import FileUploadRequest
 from app.crud.create_file_metadata import create_file_metadata
 from app.core.config import settings   # 환경설정 객체 import
-from app.
 from app.database import get_db
 import boto3
