+++ src/app/schemas/file_schemas.py
@@ -188,17 +188,6 @@ class FileUploadError(BaseModel):
 
 
 
-class ForgotPasswordRequest(BaseModel):
-    username: str = Field(..., description = "비밀번호 재설정 대상 username")
 
-class ForgotPasswordResponse(BaseModel):
-    destination: Optional[str] = None
-    delivery_medium: Optional[str] = None
-    attritubute_name: Optional[str] = None
-
-class ConfrimForgotPasswordRequest(BaseModel):
-    username: str = Field(..., description = "비밀번호 재설정 대상 username")
-    confirmation_code: str = Field(..., description = "사용자에게 발송된 확인 코드")
-    new_password: str = Field(..., min_length=8, description = "새로운 비밀번호 (최소 8자 이상)")
